Database population through CSV file. Unit tests and PricesController Integration Test

+++ src/main/java/com/tangorabox/prices/infrastructure/repository/PriceRepositoryJPA.java
@@ -8,7 +8,7 @@ import reactor.core.publisher.Flux;
 import java.time.LocalDateTime;
 
 @Repository
-public interface PriceRepositoryJPA extends ReactiveCrudRepository<PriceEntity, Integer> {
+public interface PriceRepositoryJPA extends ReactiveCrudRepository<PriceEntity, Long> {
 
     Flux<PriceEntity> findAllByProductIdAndBrandIdAndStartDateLessThanEqualAndEndDateGreaterThanEqual(long productId, long brandId, LocalDateTime date, LocalDateTime date2);
